requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
libgcc_mutex=0.1=main
pytorch_select=0.2=gpu_0
tflow_select=2.3.0=mkl
absl-py=0.11.0=py36h06a4308_0
astunparse=1.6.3=py_0
blas=1.0=mkl
blinker=1.4=py36_0
brotlipy=0.7.0=py36h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
certifi=2020.11.8=py36h06a4308_0
cffi=1.14.4=py36h261ae71_0
chardet=3.0.4=py36h06a4308_1003
click=7.1.2=py_0
cryptography=3.2.1=py36h3c74f83_1
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
ffmpeg-python=0.2.0=pypi_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gast=0.3.3=py_0
google-auth=1.21.3=py_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py36hf8bcb03_0
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
keras-preprocessing=1.1.0=py_1
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=py36h06a4308_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py36hff7bd54_0
numpy=1.16.1=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py36_0
opencv-python=4.0.0.21=pypi_0
openssl=1.1.1h=h7b6447c_0
opt_einsum=3.1.0=py_0
pandas=1.1.3=py36he6710b0_0
pillow=8.0.1=py36he98fc37_0
pip=20.3=py36h06a4308_0
protobuf=3.13.0.1=py36he6710b0_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=1.7.1=py36_0
pyopenssl=20.0.0=pyhd3eb1b0_1
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
pytorch=1.3.1=cuda100py36h53c1284_0
pytz=2020.4=pyhd3eb1b0_0
readline=8.0=h7b6447c_0
requests=2.25.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.5.2=py36h0b6359f_0
setuptools=50.3.2=py36h06a4308_2
six=1.15.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.2.0=mkl_py36h5a57954_0
tensorflow-base=2.2.0=mkl_py36hd506778_0
tensorflow-estimator=2.2.0=pyh208ff02_0
termcolor=1.1.0=py36_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.4.2=cuda100py36hecfc37a_0
urllib3=1.25.11=py_0
werkzeug=1.0.1=py_0
wheel=0.36.0=pyhd3eb1b0_0
wrapt=1.12.1=py36h7b6447c_1
xz=5.2.5=h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
